requirements.txt
tensorflow==2.0.0
absl-py
Pillow